@@ -5,9 +5,7 @@ import 'package:flutter_linkify/flutter_linkify.dart';
 import '../theme/colors.dart';
 
 class DescriptionWidget extends HookWidget {
-  const DescriptionWidget({Key? key, required String text})
-      : _text = text,
-        super(key: key);
+  const DescriptionWidget({super.key, required String text}) : _text = text;
 
   final String _text;
 
