@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 
 class RulesActions extends StatelessWidget {
   const RulesActions({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
