@@ -7,8 +7,8 @@ import 'package:get/get.dart';
 
 class RulesBasic extends GetView<RulesEditController> {
   const RulesBasic({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
